refactor: split files command god module
Break the 2,558-line files command module into focused submodules for listing, workspace path operations, attachments, clipboard integration, trash behavior, platform shell helpers, watcher setup, shared types, and regression tests.

Keep the existing Tauri command names and frontend payload contracts intact while registering the command functions through their new module paths so tauri::generate_handler can still find the generated command helpers.

Centralize clipboard entrypoints behind one file-path copy function and one image-bytes copy function, leaving platform-specific implementations private to the clipboard module.

Validation: ./scripts/stage-cli-sidecar.sh --profile debug; cargo fmt --all --check; cargo test -p claudette-tauri files:: --no-default-features --features devtools,server,voice,alternative-backends,pi-sdk; RUSTFLAGS=-Dwarnings cargo test -p claudette-tauri files:: --no-default-features --features devtools,server,voice,alternative-backends,pi-sdk.
